Fix ps_loadavg on 32 bit Windows
This is a workaround for probably a mingw bug.
On older mingw, which is used for R 3.6 and before,
you get a link error for the `PdhAddEnglishCounterW()`
call. It only happens for 32 bit Windows, and it is
apparently fixed for the compiler in Rtools4.
(If it is a mingw issue at all and not something else.)

So we work around this by using `PdhAddCounter()` instead.
However, this expects localized counter names, which differ
on different systems, and we need to look up the names in
the registry first.

Hopefully solves #105, to be tested on non-English
Windows.